requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.7.0
djangorestframework==3.17.1
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.2
sqlparse==0.5.5
